requirements.txt
# Machine Learning
scikit-learn==1.0.2


# API
fastapi         # API framework
uvicorn         # web server

# MODEL
mlflow~=1.26.1
python-dotenv
